move RowOp::debug_fmt
